Changed place of equals - (Task #512)
---
Task #512: Add missing constructors for BeanProperties
